Move EnergyComponent to dedicated module
