+++ vivarium/environment/components/eco_evo/__init__.py
@@ -1,7 +1,8 @@
+from .energy.component import EnergyComponent
 from .consumption.component import ConsumptionComponent
 from .consumption.controller import ConsumptionController
 from .consumption.interface import ConsumptionInterface
 from .spawn.component import SpawnComponent
 from .spawn.controller import SpawnController
 from .spawn.interface import SpawnInterface
-from .component import EnergyComponent, ReproductionComponent
+from .component import ReproductionComponent
